+++ lib/esm/array.d.ts
@@ -29,14 +29,14 @@ export declare function filterBy<T, V>(array: Array<T>, key: keyof T, value: V, 
  * // returns [ { name: 'Pedro' }, { name: 'Marcos' }, { name: 'Maria' }, { name: 'Davi' } ]
  * editAt(1, { name: 'Marcos' }, [{ name: 'Pedro' }, { name: 'João' }, { name: 'Maria' }, { name: 'Davi' }]);
  */
-export declare function editAt<T>(index: number, value: T | Partial<T>, array: Array<T>): Partial<T>[];
+export declare function editAt<T>(index: number, value: T | ((value: T) => T), array: Array<T>): T[];
 /**
  * Edit array item by predicate
  * @example
  * // returns [ { name: 'Pedro' }, { name: 'Marcos' }, { name: 'Maria' }, { name: 'Davi' } ]
  * editWhere(p => p.name === 'Pedro', { name: 'Marcos' }, [{ name: 'Pedro' }, { name: 'João' }, { name: 'Maria' }, { name: 'Davi' }]);
  */
-export declare function editWhere<T>(predicate: (item: T) => boolean, value: T | Partial<T> | ((item: T) => Partial<T>), array: Array<T>): Partial<T>[];
+export declare function editWhere<T>(predicate: (item: T) => boolean, value: T | ((value: T) => T), array: Array<T>): T[];
 /**
  * Remove array item by index
  * @example
@@ -61,3 +61,17 @@ export declare function removeWhere<T>(predicate: (item: T) => boolean, array: A
  * searchedItems('Banana', ['banana', 'maca', 'laranja', 'abacate']);
  */
 export declare function searchedItems<T>(search: string, items: T[], fields?: (keyof T)[]): T[];
+/**
+ * Insert array item by index
+ * @example
+ * // returns [ { name: 'Pedro' }, { name: 'João' }, { name: 'Maria' }, { name: 'Davi' } ]
+ * insertAt(1, { name: 'João' }, [{ name: 'Pedro' }, { name: 'Maria' }, { name: 'Davi' }]);
+ */
+export declare function insertAt<T>(index: number, value: T, array: Array<T>): T[];
+/**
+ * Insert array item
+ * @example
+ * // returns [ 1, 2, 3, 4, 5 ]
+ * insert(5, [1, 2, 3, 4]);
+ */
+export declare function insert<T>(value: T, array: Array<T>): T[];